refactor: Moved MobileNav compontent to navigation directory in component file
Relates #67
